#20986 able to send TTL from user to deep cache layer
